core: Snapshot should not be update after rollback.
Once createAllSnapshotCommand execution encountered a rollback,
The new snapshot is deleted as part of the compensation scenario.

That will cause the end method to get null value for the snapshotId and
will cause an sql exception when it will be updated.

The update phase should be only when the tasks failed but
also when command finished successfully and the snapshot was already
created in the DB.

Bug-Url: https://bugzilla.redhat.com/917719

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/CreateAllSnapshotsFromVmCommand.java
@@ -254,8 +254,10 @@ public class CreateAllSnapshotsFromVmCommand<T extends CreateAllSnapshotsFromVmP
      *            The snapshot ID to return to being active.
      */
     protected void revertToActiveSnapshot(Guid createdSnapshotId) {
-        getSnapshotDao().remove(createdSnapshotId);
-        getSnapshotDao().updateId(getSnapshotDao().getId(getVmId(), SnapshotType.ACTIVE), createdSnapshotId);
+        if (createdSnapshotId != null) {
+            getSnapshotDao().remove(createdSnapshotId);
+            getSnapshotDao().updateId(getSnapshotDao().getId(getVmId(), SnapshotType.ACTIVE), createdSnapshotId);
+        }
         setSucceeded(false);
     }
